instantiate a student

+++ src/org/launchcode/java/demos/lsn3classes1/Student.java
@@ -31,7 +31,7 @@ public class Student {
     public double getGpa() {
         return gpa;
     }
-    public void setGpa(int aGpa) {
+    public void setGpa(double aGpa) {
         gpa = aGpa;
     }
 }
